Refactor: 조회수 타입 변경 int -> Long

+++ src/main/java/com/ll/tourdemonde/post/entity/Post.java
@@ -22,7 +22,7 @@ public class Post extends BaseTime {
 
     private String title;
     private String category;
-    private int view;
+    private Long view;
     @ManyToOne
     private Member author;
     @OneToMany(mappedBy = "post", cascade = CascadeType.ALL, orphanRemoval = true)
